Create tmx2tbx-basic-3-fr.py
This file is an adaptation of the original txm2tbx file in order to generate a TBX file for the version 3 of the TBX standard, Basic variant. It also adds the possibility to add a title and a note under the <tbxHeader> element. Note that comments are in French. It is not dedicated to kde l10n tools.